+++ c/str/py_str.cc
@@ -22,37 +22,38 @@
 namespace py {
 
 
-  static PKArgs split_into_nhot_args(1, 0, 1, false, false, {"col", "sep"});
+static PKArgs split_into_nhot_args(1, 0, 1, false, false, {"col", "sep"},
+                                   "split_into_nhot", "");
 
-  static py::oobj split_into_nhot(const PKArgs& args) {
-    DataTable* dt = args[0].to_frame();
-    std::string sep = args[1]? args[1].to_string() : ",";
+static py::oobj split_into_nhot(const PKArgs& args) {
+  DataTable* dt = args[0].to_frame();
+  std::string sep = args[1]? args[1].to_string() : ",";
 
-    Column* col0 = dt->ncols == 1? dt->columns[0] : nullptr;
-    if (!col0) {
-      throw ValueError() << "Function split_into_nhot() may only be applied to "
-        "a single-column Frame of type string;" << " got frame with "
-        << dt->ncols << " columns";
-    }
-    SType st = col0->stype();
-    if (!(st == SType::STR32 || st == SType::STR64)) {
-      throw TypeError() << "Function split_into_nhot() may only be applied to "
-        "a single-column Frame of type string;" << " received a column of type "
-        << info(st).name();
-    }
-    if (sep.size() != 1) {
-      throw ValueError() << "Parameter `sep` in split_into_nhot() must be a "
-        "single character; got '" << sep << "'";
-    }
-
-    DataTable* res = dt::split_into_nhot(col0, sep[0]);
-    return py::Frame::from_datatable(res);
+  Column* col0 = dt->ncols == 1? dt->columns[0] : nullptr;
+  if (!col0) {
+    throw ValueError() << "Function split_into_nhot() may only be applied to "
+      "a single-column Frame of type string;" << " got frame with "
+      << dt->ncols << " columns";
   }
+  SType st = col0->stype();
+  if (!(st == SType::STR32 || st == SType::STR64)) {
+    throw TypeError() << "Function split_into_nhot() may only be applied to "
+      "a single-column Frame of type string;" << " received a column of type "
+      << info(st).name();
+  }
+  if (sep.size() != 1) {
+    throw ValueError() << "Parameter `sep` in split_into_nhot() must be a "
+      "single character; got '" << sep << "'";
+  }
+
+  DataTable* res = dt::split_into_nhot(col0, sep[0]);
+  return py::Frame::from_datatable(res);
+}
 
 }
 
 void DatatableModule::init_methods_str() {
-  add<&py::split_into_nhot, py::split_into_nhot_args>("split_into_nhot", "");
+  add<&py::split_into_nhot, py::split_into_nhot_args>();
 }